Subida de cambios para versión 2.3.2.
git-svn-id: https://svn.scae.redsara.es/svn/sim/trunk@16 15075a66-5c5d-40c6-9c28-692553124c8e

+++ PlataformaMensajeriaIOP/src/es/minhap/plataformamensajeria/iop/dao/QueryExecutorDestinatariosMensajes.java
@@ -48,7 +48,7 @@ public interface QueryExecutorDestinatariosMensajes {
 	 * @param reintentos
 	 * @return
 	 */
-	public List<EnvioGISSXMLBean> obtenerMensajesReenvioGISS(Integer servicio, Integer reintentos);
+	public List<EnvioGISSXMLBean> obtenerMensajesReenvioGISS(Long servicio, Integer reintentos);
 	
 	
 	/**
